Inputs added
Created the ability to enter how many players. Then take in that many names and if there are any alternate victory point cards in the Kingdom. Then it displays a counter for every player with the name entered. Functionality for having island and gardens does not work yet but that will be in the next commit.